+++ src/Material_TEST.cc
@@ -113,6 +113,16 @@ void MaterialTest::MaterialProperties(const std::string &_renderEngine)
   material->SetTransparency(transparency);
   EXPECT_DOUBLE_EQ(transparency, material->Transparency());
 
+  // alpha from texture
+  bool alphaFromTexture = true;
+  double alphaThreshold = 0.9;
+  bool twoSidedEnabled = false;
+  material->SetAlphaFromTexture(alphaFromTexture, alphaThreshold,
+      twoSidedEnabled);
+  EXPECT_EQ(material->TextureAlphaEnabled(), alphaFromTexture);
+  EXPECT_DOUBLE_EQ(material->AlphaThreshold(), alphaThreshold);
+  EXPECT_EQ(material->TwoSidedEnabled(), twoSidedEnabled);
+
   // reflectivity
   double reflectivity = 0.5;
   material->SetReflectivity(reflectivity);
@@ -278,6 +288,9 @@ void MaterialTest::Copy(const std::string &_renderEngine)
   math::Color emissive(0.6, 0.4, 0.2, 1.0);
   double shininess = 0.8;
   double transparency = 0.3;
+  bool alphaFromTexture = true;
+  double alphaThreshold = 0.9;
+  bool twoSidedEnabled = false;
   double reflectivity = 0.5;
   bool castShadows = false;
   bool receiveShadows = false;
@@ -303,6 +316,8 @@ void MaterialTest::Copy(const std::string &_renderEngine)
   material->SetEmissive(emissive);
   material->SetShininess(shininess);
   material->SetTransparency(transparency);
+  material->SetAlphaFromTexture(alphaFromTexture, alphaThreshold,
+      twoSidedEnabled);
   material->SetReflectivity(reflectivity);
   material->SetCastShadows(castShadows);
   material->SetReceiveShadows(receiveShadows);
@@ -329,6 +344,9 @@ void MaterialTest::Copy(const std::string &_renderEngine)
   EXPECT_EQ(emissive, clone->Emissive());
   EXPECT_DOUBLE_EQ(shininess, clone->Shininess());
   EXPECT_DOUBLE_EQ(transparency, clone->Transparency());
+  EXPECT_EQ(alphaFromTexture, clone->TextureAlphaEnabled());
+  EXPECT_DOUBLE_EQ(alphaThreshold, clone->AlphaThreshold());
+  EXPECT_EQ(twoSidedEnabled, clone->TwoSidedEnabled());
   EXPECT_DOUBLE_EQ(reflectivity, clone->Reflectivity());
   EXPECT_EQ(castShadows, clone->CastShadows());
   EXPECT_EQ(receiveShadows, clone->ReceiveShadows());
@@ -361,6 +379,9 @@ void MaterialTest::Copy(const std::string &_renderEngine)
   EXPECT_EQ(emissive, copy->Emissive());
   EXPECT_DOUBLE_EQ(shininess, copy->Shininess());
   EXPECT_DOUBLE_EQ(transparency, copy->Transparency());
+  EXPECT_EQ(alphaFromTexture, copy->TextureAlphaEnabled());
+  EXPECT_DOUBLE_EQ(alphaThreshold, copy->AlphaThreshold());
+  EXPECT_EQ(twoSidedEnabled, copy->TwoSidedEnabled());
   EXPECT_DOUBLE_EQ(reflectivity, copy->Reflectivity());
   EXPECT_EQ(castShadows, copy->CastShadows());
   EXPECT_EQ(receiveShadows, copy->ReceiveShadows());
@@ -390,6 +411,8 @@ void MaterialTest::Copy(const std::string &_renderEngine)
   comMat.SetEmissive(ambient);
   comMat.SetShininess(shininess);
   comMat.SetTransparency(transparency);
+  comMat.SetAlphaFromTexture(alphaFromTexture, alphaThreshold,
+      twoSidedEnabled);
   comMat.SetLighting(lightingEnabled);
   comMat.SetTextureImage(textureName);
 
@@ -402,6 +425,9 @@ void MaterialTest::Copy(const std::string &_renderEngine)
   EXPECT_EQ(emissive, comCopy->Emissive());
   EXPECT_DOUBLE_EQ(shininess, comCopy->Shininess());
   EXPECT_DOUBLE_EQ(transparency, comCopy->Transparency());
+  EXPECT_EQ(alphaFromTexture, comCopy->TextureAlphaEnabled());
+  EXPECT_DOUBLE_EQ(alphaThreshold, comCopy->AlphaThreshold());
+  EXPECT_EQ(twoSidedEnabled, comCopy->TwoSidedEnabled());
   EXPECT_DOUBLE_EQ(reflectivity, comCopy->Reflectivity());
   EXPECT_EQ(lightingEnabled, comCopy->LightingEnabled());
   EXPECT_EQ(textureName, comCopy->Texture());